General app improvments
Implemented the timer function. This will allow us to check for new data
which we will need to setup the win streak. Small updates to README.
Refactored some of the code in Program and Data. App now takes in the
username and platform then passes it to the initial GetData and the
Timer class for new data.